requirements.txt
numpy==1.18.5
Pillow==7.1.2
scikit-learn==0.23.1
scipy>=1.4.1
torch==1.6.0
torchvision==0.7.0
omegaconf==2.0.6
hydra-core==1.0.0rc2
faiss-gpu==1.6.3
gdown>=3.12.2
